apply feature flag to env
