+++ enterprise/dev/ci/internal/buildkite/buildkite.go
@@ -10,35 +10,14 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
-	"math/rand"
-	"os"
 	"strconv"
 	"strings"
-	"time"
 
 	"github.com/ghodss/yaml"
 	"github.com/grafana/regexp"
 
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
-
-type featureFlags struct {
-	// StatelessBuild triggers a stateless build by overriding the default queue to send the build on the stateless
-	// agents and forces a MainDryRun type build to avoid impacting normal builds.
-	//
-	// It is meant to test the stateless builds without any side effects.
-	StatelessBuild bool
-}
-
-// FeatureFlags are for experimenting with CI pipeline features. Use sparingly!
-var FeatureFlags = featureFlags{
-	StatelessBuild: os.Getenv("CI_FEATURE_FLAG_STATELESS") == "true" ||
-		// Always process retries on stateless agents.
-		// TODO: remove when we switch over entirely to stateless agents
-		os.Getenv("BUILDKITE_REBUILT_FROM_BUILD_NUMBER") != "" ||
-		// Roll out to 10% of builds
-		rand.NewSource(time.Now().UnixNano()).Int63()%100 < 10 || true, // TODO remove
-}
 
 type Pipeline struct {
 	Env    map[string]string `json:"env,omitempty"`
